Fix NXOS agent request chunking
math.ceil() needs to be applied to the result of the division and not
at len(entry). With this, length 0 maps to 0, 1 to 19 maps to 1 and 20
to 29 maps to 2 and so on.

+++ networking_ccloud/ml2/agent/nxos/switch.py
@@ -855,7 +855,7 @@ class NXOSSwitch(SwitchBase):
                         continue
 
                     # chunk the request
-                    for n in range(int(math.ceil(len(entry)) / self.GNMI_CHUNK_SIZE)):
+                    for n in range(math.ceil(len(entry) / self.GNMI_CHUNK_SIZE)):
                         part_entry = entry[n * self.GNMI_CHUNK_SIZE: (n + 1) * self.GNMI_CHUNK_SIZE]
                         upd = {op: part_entry}
                         self.api.set(**upd)
